+++ scripts/shared.js
@@ -4,6 +4,7 @@ import {
   decorateBlocks,
   decorateButtons,
   decorateIcons,
+  loadBlocks,
 } from './lib-franklin.js';
 
 /**
@@ -105,4 +106,20 @@ export function decorateMain(main) {
   buildAutoBlocks(main);
   decorateSections(main);
   decorateBlocks(main);
+}
+
+/**
+ * Fetch fragment by path
+ */
+export async function fetchFragment(path) {
+  const resp = await fetch(`${path}.plain.html`);
+  if (resp.ok) {
+    const container = document.createElement('main');
+    container.innerHTML = await resp.text();
+    // eslint-disable-next-line no-use-before-define
+    decorateMain(container);
+    await loadBlocks(container);
+    return container.querySelector(':scope .section');
+  }
+  return null;
 }
